+discretizations: - fixed missing conversion from int64 to double for variables lbs_q_signal and ubs_q_signal in method DFT of class signal_matrix
+linear_transforms:
- vectorized identity transform

+physical_values:
- added methods tril (lower triangular part of matrix), triu (upper triangular part of matrix), and issymmetric (determine if matrix is symmetric or skew-symmetric) to class transparent_container

+scattering/+sequences/+setups:
- implemented novel computation of times-of-flight in method times_of_flight of class setup
- fixed erroneous error message in method compute_prefactors of class setup

+scattering/+sequences/+syntheses:
- turned off anti-aliasing in function compute_p_in

+scattering:
- fixed fail for default linear_transforms in methods adjoint, tpsf, and energy_rx of class operator_born
- renamed linear_transforms to LTs in class operator_born

+++ +discretizations/signal_matrix.m
@@ -3,7 +3,7 @@
 %
 % author: Martin F. Schiffner
 % date: 2019-03-27
-% modified: 2019-09-10
+% modified: 2019-10-25
 %
 classdef signal_matrix
 
@@ -129,9 +129,9 @@ classdef signal_matrix
 % ensure equal subclasses of physical_values.physical_quantity
             auxiliary.mustBeEqualSubclasses( 'physical_values.time', axes_t.delta );
 
-            % extract integer bounds and temporal sampling periods
-            lbs_q_signal = reshape( [ axes_t.q_lb ], size( signal_matrices ) );
-            ubs_q_signal = reshape( [ axes_t.q_ub ], size( signal_matrices ) );
+            % extract integer bounds (int64) and temporal sampling periods
+            lbs_q_signal = reshape( double( [ axes_t.q_lb ] ), size( signal_matrices ) );
+            ubs_q_signal = reshape( double( [ axes_t.q_ub ] ), size( signal_matrices ) );
             deltas = reshape( [ axes_t.delta ], size( signal_matrices ) );
 
             % ensure nonempty T_ref
